# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaterOS SaaS API Documentation",
        "description": "Dokumentasi RESTful API untuk platform SaaS Management Catering End-to-End CaterOS.",
        "contact": {
            "name": "CaterOS Engineering Team",
            "email": "dev@cateros.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tenant/menu-items": {
            "get": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Daftar Item Menu & HPP",
                "description": "Get all menu items with category and recipes BOM.",
                "operationId": "031fc802e9fdbfcfc2211b40d774287c",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar menu berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Tambah Item Menu Baru & Resep BOM",
                "description": "Store new menu item with optional initial BOM recipe.",
                "operationId": "1058f7fe1b30df7aa521b36f4500e0ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "selling_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ayam Bakar Madu Spesial"
                                    },
                                    "menu_category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "MN-AYAM-01"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ayam bakar dengan bumbu madu dan rempah pilihan"
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "portion_unit": {
                                        "type": "string",
                                        "example": "porsi"
                                    },
                                    "recipes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "raw_material_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "gram"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/menu-items/{id}": {
            "get": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Detail Item Menu & Resep BOM",
                "description": "Show menu item detail with BOM recipes.",
                "operationId": "1a34bab7d89bc6b26af2d1535378a5b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail menu berhasil diambil"
                    },
                    "404": {
                        "description": "Menu tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Update Item Menu & Resep BOM",
                "description": "Update menu item and sync BOM recipes.",
                "operationId": "46fe5843a4b722b2dee22d1feef23dc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ayam Bakar Madu Super Pedas"
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "example": 27000
                                    },
                                    "recipes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "raw_material_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 250
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "gram"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu berhasil diperbarui"
                    },
                    "404": {
                        "description": "Menu tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Hapus Item Menu",
                "description": "Delete menu item.",
                "operationId": "e42f682b6a9670d9504a64a3ac927ac2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu berhasil dihapus"
                    },
                    "404": {
                        "description": "Menu tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/menu-items/{id}/image": {
            "post": {
                "tags": [
                    "Master Item Menu & Resep (BOM)"
                ],
                "summary": "Upload Foto Item Menu",
                "description": "Upload photo for menu item.",
                "operationId": "2552c41db4c720b55e66cb3260428c87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto berhasil diunggah"
                    },
                    "422": {
                        "description": "Validasi gambar gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer Sanctum (contoh: 1|abcdef...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "TenantHeader": {
                "type": "apiKey",
                "description": "Header ID Tenant yang sedang aktif (contoh: 1)",
                "name": "X-Tenant-ID",
                "in": "header"
            }
        }
    },
...
}